sources: refactor tar source into module.
Extract the Tar source into `sources/_tar.py`.

LP: #1648906
